Fixing publisher info
