+++ pep/types/glint.d.ts
@@ -15,12 +15,8 @@ import ModalDialog from '@gavant/glint-template-types/types/@gavant/gavant-ember
 import ModalDialogBody from '@gavant/glint-template-types/types/@gavant/gavant-ember-modals/modal-dialog/body';
 import ChangesetInput from '@gavant/glint-template-types/types/@gavant/gavant-ember-validations/changeset-input';
 import FormValidator from '@gavant/glint-template-types/types/@gavant/gavant-ember-validations/form-validator';
-import {
-    FormValidatorChildSignature
-} from '@gavant/glint-template-types/types/@gavant/gavant-ember-validations/form-validator/child';
-import InputValidator, {
-    InputValidatorSignature
-} from '@gavant/glint-template-types/types/@gavant/gavant-ember-validations/input-validator';
+import { FormValidatorChildSignature } from '@gavant/glint-template-types/types/@gavant/gavant-ember-validations/form-validator/child';
+import InputValidator, { InputValidatorSignature } from '@gavant/glint-template-types/types/@gavant/gavant-ember-validations/input-validator';
 import VerticalCollection from '@gavant/glint-template-types/types/@html-next/vertical-collection';
 import AnimatedContainer from '@gavant/glint-template-types/types/ember-animated/animated-container';
 import { AnimatedEachCurly } from '@gavant/glint-template-types/types/ember-animated/animated-each';
